pyproject.toml
[tool.poetry]
name = "test-docker-app"
version = "0.1.0"
description = "Your new great project"
authors = ["Your name"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
fastapi = {extras = ["standard"], version = "^0.115"}
uvicorn = {extras = ["standard"], version = "^0.34"}
django-split-settings = "^1.1.0"
tortoise-orm = "^0.25.0"
aerich = "^0.8"
httpx = "^0.28"
sentry-sdk = "^2.25"
pydantic-i18n = "^0.4.5"
fastapi-babel = "^1"
fastapi-pagination = "^0.12"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
pydantic = "^2.11.3"
pydantic-settings = "^2.8.1"
starlette = "^0.46.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.2"
requests = "^2.28.1"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"

[tool.aerich]
tortoise_orm = "config.tortoise_settings"
location = "app/db/migrations"
src_folder = "./app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "app",
  ".",
]
addopts = [
    "--import-mode=importlib",
]
asyncio_default_fixture_loop_scope = "module"
